address review comments, add rEADME
