--- packages/scrapers-v2/src/urls/constants.ts
@@ -1,22 +1,0 @@
-export const ALL_COLLEGES = [
-  "arts-media-design",
-  "business",
-  "computer-information-science",
-  "engineering",
-  "health-sciences",
-  "science",
-  "social-sciences-humanities",
-];
-
-/*
-Master list of old catalogs:
-https://registrar.northeastern.edu/group/academic-catalogs/
-
-Catalogs that are scrape-able (HTML):
-2021–2022 Catalog and Course Descriptions (current year)
-2020–2021 Catalog and Course Descriptions
-2019–2020 Catalog and Course Descriptions
-2018–2019 Catalog and Course Descriptions
-2017–2018 Catalog and Course Descriptions
-2016–2017 Catalog and Course Descriptions
-*/
