+++ packages/scrapers-v2/src/urls/scrape_urls.ts
@@ -1,14 +1,18 @@
-import { appendPath, loadHTML } from "../utils";
+import { convertToHierarchy, joinParts, loadHTML } from "../utils";
 import { CatalogHierarchy, College, CatalogPath } from "./types";
 
 /**
  * Scrapes all catalog entries underneath the colleges.
- * 
+ *
  * @param start starting year (must be end year - 1)
  * @param end   ending year
  * @returns     a hierarchy of catalog entry links
  */
 export const scrapeMajorLinks = async (start: number, end: number) => {
+  // technically we could specify year via 1 number, but this is ambiguous
+  // ex would "2022" refer to catalog year 2021-2022, or 2022-2023?
+  // instead, we are explicit and require two sequential numbers, like 2021-2022
+
   if (start !== end - 1) {
     throw new Error("start should == end-1");
   }
@@ -43,21 +47,17 @@ export const scrapeMajorLinksForUrl = async (
   path: string
 ): Promise<CatalogHierarchy> => {
   const paths = getPathParts(path);
-  try {
-    const initStack = Object.values(College).map((college) => ({
-      path: [...paths, college],
-    }));
-    const catalogPaths = await scrapeLinks(baseUrl, initStack);
-    return convertToHierarchy(baseUrl, catalogPaths);
-  } catch (e) {
-    throw e;
-  }
+  const initQueue = Object.values(College).map((college) => ({
+    path: [...paths, college],
+  }));
+  const catalogPaths = await scrapeLinks(baseUrl, initQueue);
+  return convertToHierarchy(baseUrl, catalogPaths);
 };
 
 /**
  *
  * Retrieves all sub-entries of the given initial queue in BFS fashion using the catalog sidebar hierarchy.
- *  
+ *
  * @param baseUrl   the base catalog URL, i.e. https://catalog.northeastern.edu
  * @param initQueue a queue of parent entries
  * @returns         a flat list of all the last level children catalog entries
@@ -86,43 +86,6 @@ const scrapeLinks = async (
   return results;
 };
 
-/**
- * Converts a flat list of entries to catalog hierarchy.
- * 
- * @param base         the base catalog URL, i.e. https://catalog.northeastern.edu
- * @param catalogPaths a flat list of paths
- * @returns            catalog hierarchy
- */
-const convertToHierarchy = (
-  base: string,
-  catalogPaths: CatalogPath[]
-): CatalogHierarchy => {
-  const hierarchy: CatalogHierarchy = {};
-  for (const { path } of catalogPaths) {
-    let obj: CatalogHierarchy = hierarchy;
-
-    // For each part of the path, add it to the hierarchy
-    for (let i = 0; i < path.length - 1; i += 1) {
-      const part = path[i];
-      if (!(part in obj)) {
-        obj[part] = {};
-      }
-      const child = obj[part];
-      if (typeof child === "string") {
-        throw new Error(
-          "Hierarchy was inconsistent: found a child, where a parent was expected"
-        );
-      }
-      obj = child;
-    }
-
-    const last = path[path.length - 1];
-    // Obj should equal the parent of the entry
-    obj[last] = joinParts(base, path).toString();
-  }
-  return hierarchy;
-};
-
 const isParent = (el: Cheerio) => {
   return el.hasClass("isparent");
 };
@@ -142,20 +105,18 @@ const getLinkForEl = (element: Cheerio): string[] => {
 };
 
 const getChildrenForPathId = ($: CheerioStatic, url: URL) => {
-  // for getElementById, forward slashes need to be escaped
+  // The catalog entries have an ID equal to the path, with a trailing slash
+  // We select the element via its ID
+  // Note: for getElementById, forward slashes need to be escaped
   const id = url.pathname.split("/").join("\\/");
   const current = $(`#${id}\\/`);
   return current.children();
 };
 
-const fetchUrlHtml = (url: URL) =>
-  new Promise<{ $: CheerioStatic; url: URL }>((res) => {
-    loadHTML(url.href).then((r) => res({ $: r, url }));
-  });
-
-const joinParts = (base: string, parts: string[]) => {
-  return appendPath(base, parts.join("/"));
-};
+const fetchUrlHtml = async (url: URL): Promise<{ $: CheerioStatic; url: URL }> => {
+  const r = await loadHTML(url.href);
+  return { $: r, url };
+}
 
 const getUrlHtmls = (queue: CatalogPath[], base: string) => {
   return queue.map(({ path }) => joinParts(base, path)).map(fetchUrlHtml);
