+++ packages/scrapers-v2/src/urls/types.ts
@@ -1,3 +1,16 @@
+/*
+Master list of old catalogs:
+https://registrar.northeastern.edu/group/academic-catalogs/
+
+Catalogs that are scrape-able (HTML):
+2021–2022 Catalog and Course Descriptions (current year)
+2020–2021 Catalog and Course Descriptions
+2019–2020 Catalog and Course Descriptions
+2018–2019 Catalog and Course Descriptions
+2017–2018 Catalog and Course Descriptions
+2016–2017 Catalog and Course Descriptions
+*/
+
 export enum College {
   ARTS_MEDIA_DESIGN = "arts-media-design",
   BUSINESS = "business",
@@ -8,14 +21,27 @@ export enum College {
   SOCIAL_SCIENCES_HUMANITIES = "social-sciences-humanities",
 }
 
+/**
+ * Represents the structure of catalog hierarchy, with each level represented
+ * by an object. The "leaf"s are the full URLs to each entry.
+ *
+ * For example, the entry
+ * /undergraduate/computer-information-science/computer-science/bscs/
+ * would be accessed by
+ * hierarchy["undergraduate"]["computer-information-science"]["computer-science"]["bscs"]
+ * and would yield the full string URL to the entry.
+ */
 export type CatalogHierarchy = {
   [part: string]: string | CatalogHierarchy;
 };
 
+/**
+ * Represents a path to a catalog entry.
+ *
+ * ex: "undergraduate", "computer-information-science"
+ * or, in the archive case:
+ * ex: "archive", "2018-2019", "undergraduate", "computer-information-science"
+ */
 export type CatalogPath = {
-  // base: https://catalog.northeastern.edu/
   path: Array<string>;
-  // ex: "undergraduate", "computer-information-science"
-  // or, in the archive case:
-  // ex: "archive", "2018-2019", "undergraduate", "computer-information-science"
 };
